# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Spy Cat Agency API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/missions/{id}/targets": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new target to an existing mission. A target cannot be added if the mission is already completed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Missions"
                ],
                "summary": "Add a target to a mission",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Target information to add",
                        "name": "target",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.AddTargetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Target added to mission successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.TargetResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid mission ID, invalid request body, or business rule violation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/spycats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all spy cats in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cats"
                ],
                "summary": "Get all spy cats",
                "responses": {
                    "200": {
                        "description": "List of spy cats retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.SpyCatListResponseDTO"
                            }
                        }
                    },
                    "404": {
                        "description": "No spy cats found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new spy cat in the system with the specified attributes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cats"
                ],
                "summary": "Create a new spy cat",
                "parameters": [
                    {
                        "description": "Spy Cat information",
                        "name": "cat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.SpyCatRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Spy cat created successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.SpyCatSingleResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error or service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove multiple spy cats from the system by their IDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cats"
                ],
                "summary": "Delete multiple spy cats",
                "parameters": [
                    {
                        "description": "Array of spy cat IDs to delete",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.DeleteIds"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Spy cats deleted successfully"
                    },
                    "400": {
                        "description": "Bad request - invalid request body or service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "No spy cats found to delete",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/spycats/{spycatId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve detailed information about a specific spy cat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cats"
                ],
                "summary": "Get a spy cat by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Spy Cat ID",
                        "name": "spycatId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Spy cat retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.SpyCatSingleResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID format or service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Spy cat not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a spy cat from the system by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cats"
                ],
                "summary": "Delete a spy cat",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Spy Cat ID",
                        "name": "spycatId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Spy cat deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID format or service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Spy cat not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the salary of an existing spy cat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cats"
                ],
                "summary": "Update spy cat salary",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Spy Cat ID",
                        "name": "spycatId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New salary information",
                        "name": "salary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.SalaryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salary updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Cat"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID format, invalid salary, or service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Spy cat not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.AddTargetRequest": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
...
        "dtos.DeleteIds": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dtos.SalaryUpdateRequest": {
            "type": "object",
            "properties": {
                "salary": {
                    "type": "number"
                }
            }
        },
        "dtos.SpyCatListResponseDTO": {
            "type": "object",
            "properties": {
                "breed": {
                    "type": "string"
                },
                "experience_years": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                }
            }
        },
        "dtos.SpyCatRequest": {
            "type": "object",
            "properties": {
                "breed": {
                    "type": "string"
                },
                "experience_years": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                }
            }
        },
        "dtos.SpyCatSingleResponseDTO": {
            "type": "object",
            "properties": {
                "breed": {
                    "type": "string"
                },
                "experience_years": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dtos.TargetResponseDTO": {
            "type": "object",
            "properties": {
                "complete_state": {
                    "$ref": "#/definitions/models.CompleteState"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mission_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Cat": {
            "type": "object",
            "properties": {
                "breed": {
                    "type": "string"
                },
                "experience_years": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.CompleteState": {
            "type": "string",
            "enum": [
                "in_progress",
                "completed"
            ],
            "x-enum-varnames": [
                "InProgress",
                "Completed"
            ]
        }
    },
...
}